third_party/WebKit/LayoutTests/plugins/npp-set-window-called-during-destruction-expected.txt

CONSOLE MESSAGE: line 12: PLUGIN: NPP_New: Could not find a test named "npp-set-window-called-during-destruction", maybe its .cpp file wasn't added to the build system?
CONSOLE ERROR: line 12: Uncaught TypeError: embed.setWillBeDestroyed is not a function

This tests whether a NPP_SetWindow is called during plugin destruction.
